test: 拆分 smoke 文件到 notification/trakt/logs/proxy
