+++ tests/api/test_logs_comprehensive.py
@@ -382,3 +382,19 @@ async def test_get_logs_grouped_disabled(app_with_auth):
                         data = response.json()
                         assert "content" in data["data"]
                         assert "groups" not in data["data"]
+
+
+# ===== 以下自 smoke 并入 =====
+
+
+@pytest.mark.asyncio
+async def test_logs_clear_truncates_file(app_with_auth, tmp_path):
+    log_f = tmp_path / "run.log"
+    log_f.write_text("line1\nline2\n", encoding="utf-8")
+
+    with patch("app.api.logs.resolved_dev_log_file_path", return_value=log_f):
+        transport = ASGITransport(app=app_with_auth)
+        async with AsyncClient(transport=transport, base_url="http://test") as ac:
+            r = await ac.post("/api/logs/clear")
+    assert r.status_code == 200
+    assert log_f.read_text(encoding="utf-8") == ""
